Update vello and parley deps...
... and everything else, apparently. This pushes the new font code changes through the ecosystem.

Brings everything into sync and should put us in a good place to start releasing.

This works as is but depends on a vello commit that is still under review: linebender/vello#425 should land first and this should be updated with the new git rev before merging.

+++ src/widget/piet_scene_helpers.rs
@@ -1,5 +1,5 @@
-use vello::kurbo::{self, Affine, Rect, Shape};
-use vello::peniko::{BrushRef, Color, ColorStopsSource, Fill, Gradient, Stroke};
+use vello::kurbo::{self, Affine, Rect, Shape, Stroke};
+use vello::peniko::{BrushRef, Color, ColorStopsSource, Fill, Gradient};
 use vello::SceneBuilder;
 
 #[derive(Debug, Clone, Copy)]
@@ -15,7 +15,7 @@ pub fn stroke<'b>(
     stroke_width: f64,
 ) {
     builder.stroke(
-        &Stroke::new(stroke_width as f32),
+        &Stroke::new(stroke_width),
         Affine::IDENTITY,
         brush,
         None,
